Added a protection in the TestPolicyConstruction unit test to use a team size of 32 for only the OpenMPTarget backend. Use long long int instead of int64_t in incremental test11* to be in sync with the SYCL backend.

+++ core/unit_test/incremental/Test11c_ParallelFor_ThreadVectorRange.hpp
@@ -54,8 +54,7 @@ namespace Test {
 
 template <class ExecSpace>
 struct Hierarchical_ForLoop_C {
-  void run(const int64_t pN, const int64_t sX, const int64_t sY,
-           const int64_t sZ) {
+  void run(const int pN, const int sX, const int sY, const int sZ) {
     using team_policy = Kokkos::TeamPolicy<ExecSpace>;
     using member_type = typename Kokkos::TeamPolicy<ExecSpace>::member_type;
